requirements.txt
cffi==1.16.0
clr-loader==0.2.6
numpy==2.0.0
pandas==2.2.2
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythonnet==3.0.3
pytz==2024.1
six==1.16.0
tzdata==2024.1
